add property hover with mock data based on @code-ui/hover
