+++ packages/app-schema-editor/by-selection-state/selection-instance-component.tsx
@@ -1,17 +1,12 @@
 import React, { useEffect, useState } from "react";
 import { nodes } from "@design-sdk/core";
 import { ISingleLayerProperty } from "../types";
-import { PropsInterfaceView } from "../interface-code-builder/props-interface-view";
 import { NameCases, nameit } from "@coli.codes/naming";
-import { FigmaNumber } from "@design-sdk/figma/features/variant";
 import { MappedPropertyStorage } from "../storage";
 import { CodeBox } from "@ui/codebox";
 import { buildeExampleDataDeclaration } from "../interface-code-builder";
 import this_interface_builder from "./selection-instance-component.coli";
-import {
-  reactNamer,
-  tsNamer,
-} from "../interface-code-builder/scoped-property-id-namer";
+import { tsNamer } from "../interface-code-builder/scoped-property-id-namer";
 import { stringfy } from "coli";
 import { CodeStyleWrapper } from "./_shared-components";
 
@@ -47,20 +42,7 @@ export default function (props: { node: nodes.light.IReflectNodeReference }) {
   return (
     <CodeStyleWrapper>
       <CodeBox language="typescript" code={interface_code_string} />
-      {/* <PropsInterfaceView
-        onInterfaceNameChange={() => {}}
-        properties={
-          properties?.map((i) => {
-            return {
-              key: i.schema.name,
-              type: FigmaNumber, // FIXME: change this to - i.schema.type
-              nullable: false, // TODO:
-            };
-          }) ?? []
-        }
-        initialInterfaceName={interfaceName}
-        onChange={() => {}}
-      /> */}
+
       <CodeBox
         language="jsx"
         code={buildeExampleDataDeclaration({
