+++ packages/app-schema-editor/by-selection-state/selection-master-component.tsx
@@ -1,17 +1,11 @@
 import React, { useEffect, useState } from "react";
-import { nodes, utils } from "@design-sdk/core";
-import { ISingleLayerProperty, IProperties } from "../types";
-import { PluginSdk } from "@plugin-sdk/app";
-import { PropsInterfaceView } from "../interface-code-builder/props-interface-view";
+import { nodes } from "@design-sdk/core";
+import { ISingleLayerProperty } from "../types";
 import { NameCases, nameit } from "@coli.codes/naming";
-import { FigmaNumber } from "@design-sdk/figma/features/variant";
 import { MappedPropertyStorage } from "../storage";
 import { CodeBox } from "@ui/codebox";
 import this_interface_builder from "./selection-master-component.coli";
-import {
-  reactNamer,
-  tsNamer,
-} from "../interface-code-builder/scoped-property-id-namer";
+import { tsNamer } from "../interface-code-builder/scoped-property-id-namer";
 import { stringfy } from "coli";
 import { CodeStyleWrapper } from "./_shared-components";
 export default function (props: { node: nodes.light.IReflectNodeReference }) {
@@ -43,20 +37,6 @@ export default function (props: { node: nodes.light.IReflectNodeReference }) {
   return (
     <CodeStyleWrapper>
       <CodeBox language="typescript" code={interface_code_string} />
-      {/* <PropsInterfaceView
-        onInterfaceNameChange={() => {}}
-        properties={
-          properties?.map((i) => {
-            return {
-              key: i.schema.name,
-              type: FigmaNumber, // FIXME: change this to - i.schema.type
-              nullable: false, // TODO:
-            };
-          }) ?? []
-        }
-        initialInterfaceName={interfaceName}
-        onChange={() => {}}
-      /> */}
     </CodeStyleWrapper>
   );
 }
