+++ packages/app-schema-editor/by-selection-state/selection-variant-instance.tsx
@@ -13,7 +13,7 @@ import {
   jsxViewExampleBuilder,
 } from "../interface-code-builder";
 import { nameit, NameCases } from "@coli.codes/naming";
-import { PropsInterfaceView } from "../interface-code-builder/props-interface-view";
+import { PropsInterfaceView } from "../components/props-interface-view";
 import styled from "@emotion/styled";
 import { MappedPropertyStorage } from "../storage";
 import { ISingleLayerProperty } from "../types";
