+++ packages/app-schema-editor/components/props-interface-view.tsx
