+++ packages/app-schema-editor/package.json
@@ -5,6 +5,8 @@
   "version": "0.0.0",
   "private": false,
   "dependencies": {
-    "@code-ui/interface": "^0.0.5"
+    "@code-ui/hover": "^0.0.2",
+    "@code-ui/interface": "^0.0.5",
+    "@radix-ui/react-hover-card": "^0.1.0"
   }
 }
